+++ application/views/buzones/listado.php
@@ -11,7 +11,17 @@
 	      </div>
 
     <br>
-<h1>LISTADO DE COMENTARIOS</h1>
+    <div class="row"; style="background-color: PeachPuff;">
+      <div class="col-md-8 text-center">
+         <h1>LISTADO DE COMENTARIOS</h1>
+      </div>
+			<div class="col-md-4">
+					<br>
+					<br>
+					<a href="<?php echo site_url('buzones/nuevo'); ?>"class="btn btn-primary"> <i class="glyphicon glyphicon-plus"></i>Agregar Comentario</a>
+				</div>
+    </div>
+    <br>
 <br>
 <?php if ($buzones): ?>
     <table class="table table-striped  table-bordered table-hover">
@@ -49,7 +59,7 @@
                  &nbsp; es para espacios-->
               <td class="text-center">
                   <a href="#" title="Editar Buzon"><i class="glyphicon glyphicon-pencil"> </i></a>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                  <a href="<?php echo site_url(); ?>/buzones/eliminar/<?php echo $filaTemporal->id_bu;?>" title="Eliminar Bunzon"><i class="glyphicon glyphicon-trash"style="color:red"> </i></a>
+                  <a href="<?php echo site_url(); ?>/buzones/eliminar/<?php echo $filaTemporal->id_bu;?>" title="Eliminar Bunzon"onclick="return confirm('¿Estas seguro de eliminar permanenteme el registro?');"><i class="glyphicon glyphicon-trash"style="color:red"> </i></a>
               </td>
             </tr>
